SQLserver bug with byte array
To use the test case you need a ms sql server and change the
persistence.xml. Everything is prepared, so you only need to change the
hostname if your user and table is `test_dn`.

The problem is the byte-array. In the Person object we have a byte array
field and this works normally, but in ms sqlserver the column created
has the type varbinary with a length of 1.

This length is much to small to save the simple "Test" name as byte
array and so we get the error message that the data is truncated and the
insertion of the date leads to a exception.

I think the field should be varbinary(max) instead of varbinary(1).